Search/defaultValue: first pass
Getting this to render properly seems like it takes more effort than it
may be worth. Additionally, this causes the following error several
times on re-render (like, say, when checking the 'autofocus' checkbox on
the demo):

    Warning: flattenChildren(...): Encountered two children with the
      same key, `Custom`. Child keys must be unique; when two children
      share a key, only the first child will be used.

As re-rendering will not happen as much under normal circumstances,
maybe this doesn't matter.